Combine worker event factory handlers (#416)
