+++ src/Services/WorkerEventFactory/EventHandler/EventHandlerInterface.php
@@ -10,7 +10,5 @@ use App\Event\EventInterface;
 
 interface EventHandlerInterface
 {
-    public function handles(EventInterface $event): bool;
-
     public function createForEvent(Job $job, EventInterface $event): ?WorkerEvent;
 }
